requirements.txt
llama-parse
python-dotenv
llama-index
llama-index-multi-modal-llms-ollama
llama-index-llms-anthropic 
llama-index-multi-modal-llms-anthropic
llama-index-embeddings-huggingface
llama-index-llms-gemini
llama-index-multi-modal-llms-gemini
langchain
langchain-community
langchain-ollama
faiss-cpu
chromadb
dash
pandas
dash-ag-grid
streamlit